+++ src/frontend/src/config.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = 'v2.190.1';
+export const APP_VERSION = 'v2.191.1';